Add `Timestamp` to the `OwnedDeliveryResult` tuple.
**Note:** This is an API breaking change.

Currently the `Timestamp` value is being dropped as part of the
conversion from `BorrowedMessage` to `OwnedDeliveryResult`, however this
value is updated once the message has been published and callers should
have access to the updated value. As far as I can tell, this is the
**only** other value that is updated as part of publishing.

An alternative to expanding the tuple would be to replace it with a
struct. This is a bit more disruptive now, but would making adding
values in the future a backwards compatible change. So it depends how
likely it is that there will be further new values.

This could also be done in a way that isn't a breaking change, but would
involve duplicating functions and copying some types, i.e.
`DeliveryFuture`. It doesn't seem like that level of complexity is
justified given the pre 1.0 status.